feat: 스타트 페이지 생성
